# btf: add support for marshaling Type and use it for Program and Map