… btf/btf_test.go
… btf/core_reloc_test.go
… btf/marshal_test.go
… btf/strings_test.go
… btf/traversal.go
… btf/traversal_test.go
… btf/types_test.go
… collection.go
… collection_test.go
… elf_reader.go
… elf_reader_test.go
… internal/deque.go
… linker_test.go
… map.go
… map_test.go
… prog.go
… testdata/btf_map_init-eb.elf
… testdata/btf_map_init-el.elf
… testdata/btf_map_init.c
… types.go