@lmb 2022-10-17
@ti-mo based on our review I did the following:
- Unexported Encoder and added MarshalMapKV instead. Since encoder is now not exported I didn't bother renaming encoder.Encode, we can do that later / never ;P
- I reworked encoder.Add to be less confusing. I think I found a middle ground between your idea and my initial implementation. The pending queue is now created directly in `Add()` which is a lot clearer, thanks for prodding me to do this.
- The coverage for `encoder` is OK, only declTag and typeTag are not covered.