@lmb 2022-10-24
> Is there more related follow-up work that needs to go in after this?

I have some optimizations for `map[Type]T` in a branch, but the code in the PR can stand on it's own.